fg keymaps renamed to lg so it does not interfere with vim f key and/or plugins such as eyeliner.lua
It makes more sense to keep keymaps under [l]aravel [*] for the
plugin if it has to set them.

f is a vim command for find, and having a keybinding fg delays the
trigger of just f, but also makes it impossible to use fg to find the
next occurence of the letter g.